user and task_status deletion restriction

+++ src/main/java/hexlet/code/service/TaskStatusService.java
@@ -5,8 +5,10 @@ import hexlet.code.dto.task_status.TaskStatusResponseDto;
 import hexlet.code.dto.task_status.TaskStatusUpdateDto;
 import hexlet.code.exception.AlreadyExistException;
 import hexlet.code.exception.NotFoundException;
+import hexlet.code.exception.RestrictionException;
 import hexlet.code.mapper.TaskStatusMapper;
 import hexlet.code.model.TaskStatus;
+import hexlet.code.repository.TaskRepository;
 import hexlet.code.repository.TaskStatusRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
@@ -22,6 +24,7 @@ import java.util.stream.Collectors;
 public class TaskStatusService {
 
     private final TaskStatusRepository taskStatusRepository;
+    private final TaskRepository taskRepository;
     private final TaskStatusMapper taskStatusMapper;
 
     public TaskStatusResponseDto getTaskStatusById(Long id) {
@@ -58,6 +61,9 @@ public class TaskStatusService {
 
     @Transactional
     public void deleteTaskStatus(Long id) {
+        if (taskRepository.existsByTaskStatusId(id)) {
+            throw new RestrictionException("The task_status cannot be deleted because it's applied to a task");
+        }
         taskStatusRepository.deleteById(id);
     }
 }
